move test related functions to command file
