+++ src/storage/txn/commands/check_secondary_locks.rs
@@ -149,3 +149,178 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for CheckSecondaryLocks {
         })
     }
 }
+
+#[cfg(test)]
+pub mod tests {
+    use super::*;
+    use crate::storage::concurrency_manager::ConcurrencyManager;
+    use crate::storage::kv::TestEngineBuilder;
+    use crate::storage::lock_manager::DummyLockManager;
+    use crate::storage::mvcc::tests::*;
+    use crate::storage::txn::commands::WriteCommand;
+    use crate::storage::Engine;
+    use kvproto::kvrpcpb::Context;
+
+    pub fn must_success<E: Engine>(
+        engine: &E,
+        key: &[u8],
+        lock_ts: impl Into<TimeStamp>,
+        expect_status: SecondaryLocksStatus,
+    ) {
+        let ctx = Context::default();
+        let snapshot = engine.snapshot(&ctx).unwrap();
+        let lock_ts = lock_ts.into();
+        let cm = ConcurrencyManager::new(lock_ts);
+        let command = crate::storage::txn::commands::CheckSecondaryLocks {
+            ctx: ctx.clone(),
+            keys: vec![Key::from_raw(key)],
+            start_ts: lock_ts,
+        };
+        let result = command
+            .process_write(
+                snapshot,
+                WriteContext {
+                    lock_mgr: &DummyLockManager,
+                    concurrency_manager: cm,
+                    extra_op: Default::default(),
+                    statistics: &mut Default::default(),
+                    pipelined_pessimistic_lock: false,
+                },
+            )
+            .unwrap();
+        if let ProcessResult::SecondaryLocksStatus { status } = result.pr {
+            assert_eq!(status, expect_status);
+            write(engine, &ctx, result.to_be_write.modifies);
+        } else {
+            unreachable!();
+        }
+    }
+
+    #[test]
+    fn test_check_async_commit_secondary_locks() {
+        let engine = TestEngineBuilder::new().build().unwrap();
+        let ctx = Context::default();
+        let cm = ConcurrencyManager::new(1.into());
+
+        let check_secondary = |key, ts| {
+            let snapshot = engine.snapshot(&ctx).unwrap();
+            let key = Key::from_raw(key);
+            let ts = TimeStamp::new(ts);
+            let command = crate::storage::txn::commands::CheckSecondaryLocks {
+                ctx: Default::default(),
+                keys: vec![key],
+                start_ts: ts,
+            };
+            let result = command
+                .process_write(
+                    snapshot,
+                    WriteContext {
+                        lock_mgr: &DummyLockManager,
+                        concurrency_manager: cm.clone(),
+                        extra_op: Default::default(),
+                        statistics: &mut Default::default(),
+                        pipelined_pessimistic_lock: false,
+                    },
+                )
+                .unwrap();
+            if !result.to_be_write.modifies.is_empty() {
+                engine.write(&ctx, result.to_be_write).unwrap();
+            }
+            if let ProcessResult::SecondaryLocksStatus { status } = result.pr {
+                status
+            } else {
+                unreachable!();
+            }
+        };
+
+        must_prewrite_lock(&engine, b"k1", b"key", 1);
+        must_commit(&engine, b"k1", 1, 3);
+        must_rollback(&engine, b"k1", 5);
+        must_prewrite_lock(&engine, b"k1", b"key", 7);
+        must_commit(&engine, b"k1", 7, 9);
+
+        // Lock CF has no lock
+        //
+        // LOCK CF       | WRITE CF
+        // --------------+---------------------
+        //               | 9: start_ts = 7
+        //               | 5: rollback
+        //               | 3: start_ts = 1
+
+        assert_eq!(
+            check_secondary(b"k1", 7),
+            SecondaryLocksStatus::Committed(9.into())
+        );
+        must_get_commit_ts(&engine, b"k1", 7, 9);
+        assert_eq!(check_secondary(b"k1", 5), SecondaryLocksStatus::RolledBack);
+        must_get_rollback_ts(&engine, b"k1", 5);
+        assert_eq!(
+            check_secondary(b"k1", 1),
+            SecondaryLocksStatus::Committed(3.into())
+        );
+        must_get_commit_ts(&engine, b"k1", 1, 3);
+        assert_eq!(check_secondary(b"k1", 6), SecondaryLocksStatus::RolledBack);
+        must_get_rollback_protected(&engine, b"k1", 6, true);
+
+        // ----------------------------
+
+        must_acquire_pessimistic_lock(&engine, b"k1", b"key", 11, 11);
+
+        // Lock CF has a pessimistic lock
+        //
+        // LOCK CF       | WRITE CF
+        // ------------------------------------
+        // ts = 11 (pes) | 9: start_ts = 7
+        //               | 5: rollback
+        //               | 3: start_ts = 1
+
+        let status = check_secondary(b"k1", 11);
+        assert_eq!(status, SecondaryLocksStatus::RolledBack);
+        must_get_rollback_protected(&engine, b"k1", 11, true);
+
+        // ----------------------------
+
+        must_prewrite_lock(&engine, b"k1", b"key", 13);
+
+        // Lock CF has an optimistic lock
+        //
+        // LOCK CF       | WRITE CF
+        // ------------------------------------
+        // ts = 13 (opt) | 11: rollback
+        //               |  9: start_ts = 7
+        //               |  5: rollback
+        //               |  3: start_ts = 1
+
+        match check_secondary(b"k1", 13) {
+            SecondaryLocksStatus::Locked(_) => {}
+            res => panic!("unexpected lock status: {:?}", res),
+        }
+        must_locked(&engine, b"k1", 13);
+
+        // ----------------------------
+
+        must_commit(&engine, b"k1", 13, 15);
+
+        // Lock CF has an optimistic lock
+        //
+        // LOCK CF       | WRITE CF
+        // ------------------------------------
+        //               | 15: start_ts = 13
+        //               | 11: rollback
+        //               |  9: start_ts = 7
+        //               |  5: rollback
+        //               |  3: start_ts = 1
+
+        match check_secondary(b"k1", 14) {
+            SecondaryLocksStatus::RolledBack => {}
+            res => panic!("unexpected lock status: {:?}", res),
+        }
+        must_get_rollback_protected(&engine, b"k1", 14, true);
+
+        match check_secondary(b"k1", 15) {
+            SecondaryLocksStatus::RolledBack => {}
+            res => panic!("unexpected lock status: {:?}", res),
+        }
+        must_get_overlapped_rollback(&engine, b"k1", 15, 13, WriteType::Lock);
+    }
+}
