+++ src/storage/txn/commands/pessimistic_rollback.rs
@@ -87,3 +87,103 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for PessimisticRollback {
         })
     }
 }
+
+#[cfg(test)]
+pub mod tests {
+    use super::*;
+    use crate::storage::kv::Engine;
+    use crate::storage::lock_manager::DummyLockManager;
+    use crate::storage::mvcc::tests::*;
+    use crate::storage::txn::commands::{WriteCommand, WriteContext};
+    use crate::storage::{concurrency_manager::ConcurrencyManager, TestEngineBuilder};
+    use kvproto::kvrpcpb::Context;
+    use txn_types::Key;
+
+    pub fn must_success<E: Engine>(
+        engine: &E,
+        key: &[u8],
+        start_ts: impl Into<TimeStamp>,
+        for_update_ts: impl Into<TimeStamp>,
+    ) {
+        let ctx = Context::default();
+        let snapshot = engine.snapshot(&ctx).unwrap();
+        let for_update_ts = for_update_ts.into();
+        let cm = ConcurrencyManager::new(for_update_ts);
+        let start_ts = start_ts.into();
+        let command = crate::storage::txn::commands::PessimisticRollback {
+            ctx: ctx.clone(),
+            keys: vec![Key::from_raw(key)],
+            start_ts,
+            for_update_ts,
+        };
+        let lock_mgr = DummyLockManager;
+        let write_context = WriteContext {
+            lock_mgr: &lock_mgr,
+            concurrency_manager: cm,
+            extra_op: Default::default(),
+            statistics: &mut Default::default(),
+            pipelined_pessimistic_lock: false,
+        };
+        let result = command.process_write(snapshot, write_context).unwrap();
+        write(engine, &ctx, result.to_be_write.modifies);
+    }
+
+    #[test]
+    fn test_pessimistic_rollback() {
+        let engine = TestEngineBuilder::new().build().unwrap();
+
+        let k = b"k1";
+        let v = b"v1";
+
+        // Normal
+        must_acquire_pessimistic_lock(&engine, k, k, 1, 1);
+        must_pessimistic_locked(&engine, k, 1, 1);
+        must_success(&engine, k, 1, 1);
+        must_unlocked(&engine, k);
+        must_get_commit_ts_none(&engine, k, 1);
+        // Pessimistic rollback is idempotent
+        must_success(&engine, k, 1, 1);
+        must_unlocked(&engine, k);
+        must_get_commit_ts_none(&engine, k, 1);
+
+        // Succeed if the lock doesn't exist.
+        must_success(&engine, k, 2, 2);
+
+        // Do nothing if meets other transaction's pessimistic lock
+        must_acquire_pessimistic_lock(&engine, k, k, 2, 3);
+        must_success(&engine, k, 1, 1);
+        must_success(&engine, k, 1, 2);
+        must_success(&engine, k, 1, 3);
+        must_success(&engine, k, 1, 4);
+        must_success(&engine, k, 3, 3);
+        must_success(&engine, k, 4, 4);
+
+        // Succeed if for_update_ts is larger; do nothing if for_update_ts is smaller.
+        must_pessimistic_locked(&engine, k, 2, 3);
+        must_success(&engine, k, 2, 2);
+        must_pessimistic_locked(&engine, k, 2, 3);
+        must_success(&engine, k, 2, 4);
+        must_unlocked(&engine, k);
+
+        // Do nothing if rollbacks a non-pessimistic lock.
+        must_prewrite_put(&engine, k, v, k, 3);
+        must_locked(&engine, k, 3);
+        must_success(&engine, k, 3, 3);
+        must_locked(&engine, k, 3);
+
+        // Do nothing if meets other transaction's optimistic lock
+        must_success(&engine, k, 2, 2);
+        must_success(&engine, k, 2, 3);
+        must_success(&engine, k, 2, 4);
+        must_success(&engine, k, 4, 4);
+        must_locked(&engine, k, 3);
+
+        // Do nothing if committed
+        must_commit(&engine, k, 3, 4);
+        must_unlocked(&engine, k);
+        must_get_commit_ts(&engine, k, 3, 4);
+        must_success(&engine, k, 3, 3);
+        must_success(&engine, k, 3, 4);
+        must_success(&engine, k, 3, 5);
+    }
+}
