+++ src/storage/txn/commands/txn_heart_beat.rs
@@ -2,9 +2,7 @@
 
 use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::LockManager;
-use crate::storage::mvcc::{
-    Error as MvccError, ErrorInner as MvccErrorInner, MvccTxn, Result as MvccResult,
-};
+use crate::storage::mvcc::{ErrorInner as MvccErrorInner, MvccTxn, Result as MvccResult};
 use crate::storage::txn::commands::{
     Command, CommandExt, TypedCommand, WriteCommand, WriteContext, WriteResult,
 };
@@ -49,10 +47,14 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for TxnHeartBeat {
             !self.ctx.get_not_fill_cache(),
             context.concurrency_manager,
         );
-        fail_point!("txn_heart_beat", |err| Err(MvccError::from(
-            crate::storage::mvcc::txn::make_txn_error(err, &self.primary_key, self.start_ts,)
-        )
-        .into()));
+        fail_point!("txn_heart_beat", |err| Err(
+            crate::storage::mvcc::Error::from(crate::storage::mvcc::txn::make_txn_error(
+                err,
+                &self.primary_key,
+                self.start_ts,
+            ))
+            .into()
+        ));
 
         let lock_ttl: MvccResult<_> =
             if let Some(mut lock) = txn.reader.load_lock(&self.primary_key)? {
@@ -112,5 +114,141 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for TxnHeartBeat {
             lock_info: None,
             lock_guards: vec![],
         })
+    }
+}
+
+#[cfg(test)]
+pub mod tests {
+    use super::*;
+    use crate::storage::concurrency_manager::ConcurrencyManager;
+    use crate::storage::kv::TestEngineBuilder;
+    use crate::storage::lock_manager::DummyLockManager;
+    use crate::storage::mvcc::tests::*;
+    use crate::storage::txn::commands::WriteCommand;
+    use crate::storage::Engine;
+    use kvproto::kvrpcpb::Context;
+
+    pub fn must_success<E: Engine>(
+        engine: &E,
+        primary_key: &[u8],
+        start_ts: impl Into<TimeStamp>,
+        advise_ttl: u64,
+        expect_ttl: u64,
+    ) {
+        let ctx = Context::default();
+        let snapshot = engine.snapshot(&ctx).unwrap();
+        let start_ts = start_ts.into();
+        let cm = ConcurrencyManager::new(start_ts);
+        let command = crate::storage::txn::commands::TxnHeartBeat {
+            ctx: Context::default(),
+            primary_key: Key::from_raw(primary_key),
+            start_ts,
+            advise_ttl,
+        };
+        let result = command
+            .process_write(
+                snapshot,
+                WriteContext {
+                    lock_mgr: &DummyLockManager,
+                    concurrency_manager: cm,
+                    extra_op: Default::default(),
+                    statistics: &mut Default::default(),
+                    pipelined_pessimistic_lock: false,
+                },
+            )
+            .unwrap();
+        if let ProcessResult::TxnStatus {
+            txn_status:
+                TxnStatus::Uncommitted {
+                    lock_ttl,
+                    min_commit_ts: _,
+                    use_async_commit: _,
+                    secondaries: _,
+                },
+        } = result.pr
+        {
+            write(engine, &ctx, result.to_be_write.modifies);
+            assert_eq!(lock_ttl, expect_ttl);
+        } else {
+            unreachable!();
+        }
+    }
+
+    pub fn must_err<E: Engine>(
+        engine: &E,
+        primary_key: &[u8],
+        start_ts: impl Into<TimeStamp>,
+        advise_ttl: u64,
+    ) {
+        let ctx = Context::default();
+        let snapshot = engine.snapshot(&ctx).unwrap();
+        let start_ts = start_ts.into();
+        let cm = ConcurrencyManager::new(start_ts);
+        let command = crate::storage::txn::commands::TxnHeartBeat {
+            ctx,
+            primary_key: Key::from_raw(primary_key),
+            start_ts,
+            advise_ttl,
+        };
+        assert!(command
+            .process_write(
+                snapshot,
+                WriteContext {
+                    lock_mgr: &DummyLockManager,
+                    concurrency_manager: cm,
+                    extra_op: Default::default(),
+                    statistics: &mut Default::default(),
+                    pipelined_pessimistic_lock: false,
+                },
+            )
+            .is_err());
+    }
+
+    #[test]
+    fn test_txn_heart_beat() {
+        let engine = TestEngineBuilder::new().build().unwrap();
+
+        let (k, v) = (b"k1", b"v1");
+
+        let test = |ts| {
+            // Do nothing if advise_ttl is less smaller than current TTL.
+            must_success(&engine, k, ts, 90, 100);
+            // Return the new TTL if the TTL when the TTL is updated.
+            must_success(&engine, k, ts, 110, 110);
+            // The lock's TTL is updated and persisted into the db.
+            must_success(&engine, k, ts, 90, 110);
+            // Heart beat another transaction's lock will lead to an error.
+            must_err(&engine, k, ts - 1, 150);
+            must_err(&engine, k, ts + 1, 150);
+            // The existing lock is not changed.
+            must_success(&engine, k, ts, 90, 110);
+        };
+
+        // No lock.
+        must_err(&engine, k, 5, 100);
+
+        // Create a lock with TTL=100.
+        // The initial TTL will be set to 0 after calling must_prewrite_put. Update it first.
+        must_prewrite_put(&engine, k, v, k, 5);
+        must_locked(&engine, k, 5);
+        must_success(&engine, k, 5, 100, 100);
+
+        test(5);
+
+        must_locked(&engine, k, 5);
+        must_commit(&engine, k, 5, 10);
+        must_unlocked(&engine, k);
+
+        // No lock.
+        must_err(&engine, k, 5, 100);
+        must_err(&engine, k, 10, 100);
+
+        must_acquire_pessimistic_lock(&engine, k, k, 8, 15);
+        must_pessimistic_locked(&engine, k, 8, 15);
+        must_success(&engine, k, 8, 100, 100);
+
+        test(8);
+
+        must_pessimistic_locked(&engine, k, 8, 15);
     }
 }
